+++ indispensable/src/jvmTest/kotlin/at/asitplus/signum/ecmath/ECCurveTest.kt
@@ -2,7 +2,7 @@ package at.asitplus.signum.indispensable.ecmath
 
 import at.asitplus.signum.indispensable.ECCurve
 import com.ionspin.kotlin.bignum.integer.toBigInteger
-import at.asitplus.test.FreeSpec
+import io.kotest.core.spec.style.FreeSpec
 import io.kotest.matchers.shouldBe
 
 /**