+++ indispensable/src/wasmJsTest/kotlin/DummyTest.kt
@@ -1,4 +1,4 @@
-import at.asitplus.test.FreeSpec
+import io.kotest.core.spec.style.FreeSpec
 
 class DummyTest: FreeSpec( {
 })