+++ supreme/src/commonTest/kotlin/at/asitplus/signum/supreme/mac/MACTest.kt
@@ -4,7 +4,7 @@ package at.asitplus.signum.supreme.mac
 import at.asitplus.signum.indispensable.HMAC
 import at.asitplus.signum.indispensable.misc.bit
 import at.asitplus.signum.supreme.b
-import at.asitplus.test.FreeSpec
+import io.kotest.core.spec.style.FreeSpec
 import io.kotest.datatest.withData
 import io.kotest.matchers.shouldBe