+++ supreme/src/iosTest/kotlin/Test.kt
@@ -1,4 +1,4 @@
-import at.asitplus.test.FreeSpec
+import io.kotest.core.spec.style.FreeSpec
 import io.kotest.matchers.shouldNotBe
 
 class ProviderTest : FreeSpec({