+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 3.0
 
 ### NEXT
+* Remove Kotest Workarounds
 
 ### 3.17.0 (Supreme 0.9.0)
 * **KDF Support**